+++ api/routers/admin/roles/list.js
@@ -8,7 +8,7 @@ module.exports = new Route({
   method: 'get',
   path: '/',
   handler: async function (ctx) {
-    const filters = queryParams.toFilters(ctx.request.query)
+    const filters = await queryParams.toFilters(ctx.request.query)
 
     var role = await Role.dataTables({
       limit: ctx.request.query.limit || 20,